pyproject.toml
[tool.poetry]
name = "chester-ml"
version = "1.0.5"
description = "A CLI tool for machine learning data providers (SQL, Mongo, local files)."
authors = ["Carlos De Silvestri <tucorreo@dominio.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "chester_ml", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "*"
scikit-learn = "*"
loguru = "*"
python-dotenv = "*"
matplotlib = "^3.10.7"
sqlalchemy = "^2.0.44"
mysql-connector-python = "^9.5.0"
pymongo = "^4.15.4"

[tool.poetry.scripts]
chester = "chester_ml.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
